refactor(web-ui): simplify file tree and remove path compression utilities
- Remove useFileTree, useFileSearch, and duplicated pathCompression modules.

- Adjust file explorer hooks and types; trim FileExplorer styles.

- Polish About dialog, session capsule, top bar, and footer actions; extend i18n.

- Trim optional Cargo.toml feature entries in desktop, relay-server, core, and webdriver.
