+++ src/apps/desktop/Cargo.toml
@@ -61,11 +61,9 @@ core-graphics = "0.23"
 dispatch = "0.2"
 objc2 = "0.6"
 objc2-foundation = "0.3"
-objc2-app-kit = "0.3"
 objc2-vision = { version = "0.3.2", features = ["VNRecognizeTextRequest", "VNRequest", "VNObservation", "VNRequestHandler", "VNUtils", "VNTypes", "objc2-core-foundation"] }
 
 [target.'cfg(windows)'.dependencies]
-win32job = { workspace = true }
 windows = { version = "0.61.3", features = [
   "Foundation",
   "Globalization",
