+++ src/apps/relay-server/Cargo.toml
@@ -34,9 +34,8 @@ tracing = "0.1"
 tracing-subscriber = { version = "0.3", features = ["env-filter"] }
 
 # Utilities
-uuid = { version = "1.0", features = ["v4", "serde"] }
 chrono = { version = "0.4", features = ["serde", "clock"] }
 dashmap = "5.5"
 rand = "0.8"
-base64 = "0.21"
+base64 = { workspace = true }
 sha2 = "0.10"
