+++ src/crates/core/Cargo.toml
@@ -64,8 +64,6 @@ grep-searcher = { workspace = true }
 grep-regex = { workspace = true }
 globset = { workspace = true }
 
-eventsource-stream = { workspace = true }
-
 # MCP Streamable HTTP client (official rust-sdk used by Codex)
 rmcp = { version = "0.12.0", default-features = false, features = [
     "auth",
