+++ src/crates/webdriver/Cargo.toml
@@ -31,7 +31,4 @@ windows = { version = "0.61.3", features = ["Win32_Foundation", "Win32_System_Co
 windows-core = "0.61.2"
 
 [target.'cfg(target_os = "linux")'.dependencies]
-glib = "0.18.5"
-gtk = "0.18.2"
 tempfile = "3"
-webkit2gtk = "2.0.2"
