+++ src/web-ui/src/app/components/AboutDialog/AboutDialog.scss
@@ -5,6 +5,30 @@
 
 @use '../../../component-library/styles/tokens' as *;
 
+// Match settings content panel: no modal chrome border on this dialog only
+.modal-overlay.bitfun-about-modal-overlay {
+  // Extra horizontal inset from viewport (左右留白)
+  padding: 24px 32px;
+
+  .modal {
+    border: none;
+    border-radius: 12px;
+    overflow: hidden;
+    background: var(--color-bg-primary);
+    box-shadow: 0 16px 40px rgba(0, 0, 0, 0.45);
+  }
+
+  .modal__header {
+    border-bottom: none;
+  }
+}
+
+@media (max-width: 480px) {
+  .modal-overlay.bitfun-about-modal-overlay {
+    padding: 16px 20px;
+  }
+}
+
 // ==================== Content area (inside Modal) ====================
 
 .bitfun-about-dialog__content {
@@ -98,44 +122,28 @@
   }
 }
 
-// Decorative dots
-.bitfun-about-dialog__dots {
-  display: flex;
-  gap: 6px;
-  margin-top: 12px;
-  position: relative;
-  z-index: 1;
-  animation: about-dots-enter 0.6s ease-out 0.25s both;
-
-  span {
-    width: 3px;
-    height: 3px;
-    background: var(--color-text-muted);
-    border-radius: 50%;
-    opacity: 0.45;
-  }
-}
-
-@keyframes about-dots-enter {
-  from { opacity: 0; }
-  to { opacity: 1; }
-}
-
 // ==================== Scrollable area ====================
 
 .bitfun-about-dialog__scrollable {
   flex: 1;
   min-height: 0;
+  display: flex;
+  flex-direction: column;
   overflow-y: auto;
   overflow-x: hidden;
-  padding: 0 24px;
-  
+  // Inset rounded panel: side margins + rounded rect (圆角矩形)
+  margin: 0 20px;
+  padding: 12px 16px 16px;
+  border-radius: 12px;
+  background: var(--color-bg-scene, var(--color-bg-primary));
 }
 
 // ==================== Info section ====================
 
 .bitfun-about-dialog__info-section {
-  margin-bottom: 16px;
+  flex-shrink: 0;
+  // 在 scrollable 剩余高度内垂直居中；内容过高时 margin 让位，仍可滚动
+  margin-block: auto;
   animation: about-section-enter 0.4s ease-out 0.2s both;
 }
 
@@ -153,7 +161,7 @@
 .bitfun-about-dialog__info-card {
   padding: 10px 14px;
   background: transparent;
-  border: 1px dashed var(--border-subtle);
+  border: none;
   border-radius: 8px;
 }
 
@@ -417,20 +425,10 @@
 
 .bitfun-about-dialog__footer {
   flex-shrink: 0;
-  padding: 16px 24px 20px;
+  padding: 16px 28px 20px;
   text-align: center;
   position: relative;
   animation: about-section-enter 0.4s ease-out 0.3s both;
-  
-  &::before {
-    content: '';
-    position: absolute;
-    top: 0;
-    left: 24px;
-    right: 24px;
-    height: 1px;
-    background: linear-gradient(90deg, transparent 0%, var(--border-subtle) 50%, transparent 100%);
-  }
 }
 
 .bitfun-about-dialog__license {
@@ -465,7 +463,9 @@
   }
 
   .bitfun-about-dialog__scrollable {
-    padding: 0 16px;
+    margin: 0 12px;
+    padding: 12px 14px 14px;
+    border-radius: 10px;
   }
 
   .bitfun-about-dialog__dependencies-grid {
@@ -473,12 +473,7 @@
   }
 
   .bitfun-about-dialog__footer {
-    padding: 14px 16px 16px;
-
-    &::before {
-      left: 16px;
-      right: 16px;
-    }
+    padding: 14px 20px 16px;
   }
 }
 
@@ -498,10 +493,6 @@
 
   .bitfun-about-dialog__divider {
     margin-top: 12px;
-  }
-
-  .bitfun-about-dialog__dots {
-    margin-top: 8px;
   }
 
   .bitfun-about-dialog__footer {
