+++ src/web-ui/src/app/components/WorkspaceFooterActions/WorkspaceFooterActions.scss
@@ -131,7 +131,8 @@
   flex-direction: row;
   align-items: stretch;
   min-width: 0;
-  min-height: 280px;
+  /* Tall enough that expanding/collapsing Agent App sublist does not resize the card; extra space flexes above Dispatcher row */
+  min-height: 400px;
   padding: 0;
   overflow: hidden;
   background: var(--color-bg-elevated, #1e1e22);
@@ -150,9 +151,11 @@
 .bitfun-nav-panel__footer-menu-col-actions {
   flex-shrink: 0;
   width: 192px;
+  min-height: 0;
   padding: $size-gap-1;
   display: flex;
   flex-direction: column;
+  align-self: stretch;
 }
 
 .bitfun-nav-panel__footer-menu-col-sep {
