+++ src/web-ui/src/app/overlay/AgenticOSWorkspace.scss
@@ -35,7 +35,6 @@
   }
 
   &--dispatcher {
-    border-radius: 0;
     border-width: 0;
     border-color: transparent;
     box-shadow: none;
