+++ src/web-ui/src/hooks/index.ts
@@ -1,5 +1,4 @@
 export * from './useModelConfigs';
 export * from './useToolExecution';
 export * from './useResizer';
-export * from './useFileSearch';
 export * from './useScrollActivity';
