+++ src/web-ui/src/tools/file-explorer/controller/ExplorerController.ts
@@ -91,7 +91,6 @@ export class ExplorerController {
   private config: ExplorerControllerConfig = {
     autoLoad: true,
     enableAutoWatch: true,
-    enablePathCompression: true,
     showHiddenFiles: false,
     sortBy: 'name',
     sortOrder: 'asc',
