+++ src/web-ui/src/tools/file-system/hooks/index.ts
@@ -1,4 +1,2 @@
 export { useFileSystem } from './useFileSystem';
-export { useFileTree } from './useFileTree';
 export type { UseFileSystemOptions, UseFileSystemReturn } from './useFileSystem';
-export type { UseFileTreeOptions, UseFileTreeReturn } from './useFileTree';
