+++ src/web-ui/src/tools/file-system/hooks/useFileSystem.ts
@@ -36,7 +36,6 @@ export function useFileSystem(options: UseFileSystemOptions = {}): UseFileSystem
     rootPath: options.rootPath,
     autoLoad: options.autoLoad ?? true,
     enableAutoWatch: options.enableAutoWatch ?? true,
-    enablePathCompression: optionOverrides.enablePathCompression ?? options.enablePathCompression ?? true,
     showHiddenFiles: optionOverrides.showHiddenFiles ?? options.showHiddenFiles ?? false,
     sortBy: optionOverrides.sortBy ?? options.sortBy ?? 'name',
     sortOrder: optionOverrides.sortOrder ?? options.sortOrder ?? 'asc',
@@ -46,7 +45,6 @@ export function useFileSystem(options: UseFileSystemOptions = {}): UseFileSystem
     options.rootPath,
     options.autoLoad,
     options.enableAutoWatch,
-    options.enablePathCompression,
     options.showHiddenFiles,
     options.sortBy,
     options.sortOrder,
